Remove reference reduce tactic
